@filipefurtad0 2021-09-08
Hey @cillian ,
Thanks for addressing this!

- Before this PR

![image](https://user-images.githubusercontent.com/49817236/132555593-d513a230-0439-4a3f-a17c-7f3d2225e599.png)

The last three runs make up the following average values:

Making averages of -> Completed 200 OK in ~ 5258ms; Views 3399ms; Active Record 908ms; Allocations 	2450543


- After this PR

![image](https://user-images.githubusercontent.com/49817236/132555843-c23dfe48-e63d-4eb1-a97d-01da853f81df.png)

Making averages of -> Completed 200 OK in ~4263 ms; Views	2823 ms; Active Record 496ms; Allocations 2236548

**Summary**

That makes up for a reduction of -> Completed 200 OK in ~19%; Views 17%; Active Record 45%; Allocations 9%

:fire: :fire: :fire: 

Ready to go!